RAG_pdf_book_mcp_Cos_Similarity_260414/pyproject.toml
[project]
name = "rag-pdf-book-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastmcp>=3.2.3",
    "numpy>=2.4.4",
    "sentence-transformers>=5.4.0",
]
